+++ fabric/src/main/java/dev/nyon/magnetic/mixins/blocks/SweetBerryBushBlockMixin.java
@@ -35,6 +35,6 @@ public class SweetBerryBushBlockMixin {
     ) {
         if (!(player instanceof ServerPlayer serverPlayer)) return true;
 
-        return MixinHelper.wrapWithConditionPlayerItemSingle(serverPlayer, itemStack);
+        return MixinHelper.wrapWithConditionPlayerItemSingle(serverPlayer, itemStack, blockPos);
     }
 }