+++ fabric/src/main/java/dev/nyon/magnetic/mixins/compat/kleeslabs/BlockBreakHandlerMixin.java
@@ -29,6 +29,6 @@ public class BlockBreakHandlerMixin {
     ) {
         if (!(event.getPlayer() instanceof ServerPlayer serverPlayer) || !(event.getLevel() instanceof ServerLevel)) return true;
         if (!(entity instanceof ItemEntity itemEntity)) return true;
-        return MixinHelper.wrapWithConditionPlayerItemSingle(serverPlayer, itemEntity.getItem());
+        return MixinHelper.wrapWithConditionPlayerItemSingle(serverPlayer, itemEntity.getItem(), event.getPos());
     }
 }